address review feedback

+++ mail-toaster.sh
@@ -836,15 +836,22 @@ unprovision_filesystems()
 	fi
 }
 
+valid_jail_name()
+{
+	case "$1" in
+		""|.|..|*[!A-Za-z0-9_.-]*) return 1 ;;
+	esac
+}
+
 unprovision_etc()
 {
-	if [ -z "$1" ] || [ -z "$MT6_ETC" ]; then return 0; fi
+	if [ -z "$MT6_ETC" ] || ! valid_jail_name "$1"; then return 0; fi
 
 	local _etc; _etc=$(get_jail_host_etc "$1")
 	if [ ! -d "$_etc" ]; then return 0; fi
 
 	tell_status "rm -r $_etc"
-	rm -r "$_etc"
+	rm -r -- "$_etc"
 }
 
 unprovision_files()
@@ -856,9 +863,9 @@ unprovision_files()
 		fi
 	done
 
-	if [ -n "$MT6_ETC" ] && [ -d "$MT6_ETC" ]; then
+	if [ -n "$MT6_ETC" ] && [ "$MT6_ETC" != "/" ] && [ -d "$MT6_ETC" ]; then
 		tell_status "rm -r $MT6_ETC"
-		rm -r "$MT6_ETC"
+		rm -r -- "$MT6_ETC"
 	fi
 
 	if grep -q "^$JAIL_NET_PREFIX" /etc/hosts; then
@@ -868,13 +875,15 @@ unprovision_files()
 
 unprovision_rc()
 {
+	if ! valid_jail_name "$1"; then return 0; fi
+
 	tell_status "disabling jail $1 startup"
 	sysrc jail_list-=" $1"
 	sysrc -f /etc/periodic.conf security_status_pkgaudit_jails-=" $1"
 
-	if [ -f /etc/jail.conf.d/$1.conf ]; then
+	if [ -f "/etc/jail.conf.d/$1.conf" ]; then
 		tell_status "deleting /etc/jail.conf.d/$1.conf"
-		rm "/etc/jail.conf.d/$1.conf"
+		rm -- "/etc/jail.conf.d/$1.conf"
 	fi
 }
